tesseract-api/requirements.txt
flask
pytesseract
requests
Pillow